Moving a Card to a separate component
