+++ src/components/Card/Card.css
@@ -1,0 +1,37 @@
+.card {
+    max-height: auto;
+    width: 100%;
+    font-size: 0.8rem;
+    background-color: #8080801a;
+    border: 2px solid #8080801a;
+    border-radius: 16px;
+    padding: 1rem;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+    .card__info {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+    }
+    .card__photo {
+        grid-area: photo;
+        img {
+            height: 110px;
+            width: 110px;
+            border-radius: 50%;
+            object-fit: contain;
+        }
+    }
+    .card__name {
+        grid-area: name;
+        font-weight: 500;
+    }
+    .card__text {
+        grid-area: text;
+        display: inline-block;
+    }
+    .card__rating {
+        grid-area: rating;
+    }
+}
