+++ src/section/About/About.css
@@ -20,7 +20,6 @@
     }
     .about__text {
         flex: 1 0 50%;
-        /* max-width: 30%; */
         h3 {
             font-size: 2rem;
             font-weight: 700;
@@ -41,44 +40,9 @@
     }
 }
 
-.card {
+.about__clients .card {
     position: absolute;
-    max-height: auto;
-    width: 100%;
     transition: all 1.5s ease-out;
-    font-size: 0.8rem;
-    background-color: #8080801a;
-    border: 2px solid #8080801a;
-    border-radius: 16px;
-    padding: 1rem;
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    .card__info {
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-    }
-    .card__photo {
-        grid-area: photo;
-        img {
-            height: 110px;
-            width: 110px;
-            border-radius: 50%;
-            object-fit: contain;
-        }
-    }
-    .card__name {
-        grid-area: name;
-        font-weight: 500;
-    }
-    .card__text {
-        grid-area: text;
-        display: inline-block;
-    }
-    .card__rating {
-        grid-area: rating;
-    }
     &:nth-child(1) {
         top: 0;
         transform: translateX(80px);
